vis_requirement.txt
open3d==0.16.0
numpy==1.22.4
matplotlib==3.6.2
pynput==1.7.6
pyyaml